--- src/components/DetentionsPage.js
@@ -1,19 +1,0 @@
-import React from 'react';
-// import PropTypes from 'prop-types';
-import Header from './Header';
-import SideBar from './SideBar';
-
-
-const DetentionsPage = () => {
-
-  return (
-    <>
-      <Header />
-      <SideBar />
-      <h1>Detentions!</h1>
-    </>
-  );
-
-};
-
-export default DetentionsPage;
