+++ src/components/HomePage/DynamicDataDisplay.css
