+++ src/components/common/Header.js
@@ -1,7 +1,7 @@
 import React from 'react';
 // import PropTypes from 'prop-types';
 import styles from './Header.css';
-import DynamicDataDisplay from '../DynamicDataDisplay';
+import DynamicDataDisplay from '../HomePage/DynamicDataDisplay';
 import useAverageDetention from '../../hooks/useAverageDetention';
