+++ src/components/common/Paging/Paging.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import usePaging from '../../hooks/use-paging';
+import usePaging from '../../../hooks/use-paging';
 // import styles from './Paging.css';
 
 const Paging = () => {
